Add random decision bursts, per-mode ticks, RSS monitoring, and effort infrastructure
CaDiCaL-style features:
- S11: Random decision bursts (sat.randec, default off) with growing
  burst length and interval scheduling
- C11: Per-mode tick tracking - swap search_ticks on stable/focused
  mode switch for accurate per-mode restart accounting
- C10: memory::get_rss() utility reads /proc/self/statm on Linux for
  actual RSS monitoring (available for GC/memory decisions)
- C1: Propagation-proportional effort tracking infrastructure and
  add_effort_budget() methods on probing/vivification (currently
  additive to existing doubling + AIMD-delay budgets)

+++ src/sat/sat_asymm_branch.cpp
@@ -758,7 +758,7 @@ namespace sat {
 
         s.m_phase = saved_phase;
         m_asymm_branch_limit *= 2;
-        if (m_asymm_branch_limit > UINT_MAX)
+        if (m_asymm_branch_limit > static_cast<int64_t>(UINT_MAX))
             m_asymm_branch_limit = UINT_MAX;
 
         CASSERT("asymm_branch", s.check_invariant());
